move wlr dependent input handler to sway
Moves input handling code that uses headers from wlroots from the common
code to sway proper, to allow for removing the wlroots dependency from
clients
